requirement.txt
pillow
lxml>=4.9.4,<5
requests-html==0.10.0
SpeechRecognition>=3.10.0
# Optional on Linux if portaudio dev headers are missing. Voice input gracefully degrades without it.
PyAudio==0.2.14
requests>=2.31.0
ollama>=0.1.0
PyWhatKit>=7.0
openai-whisper>=20240314
keyboard>=0.13.5
pydub>=0.25.1
numpy>=1.24.0
